[00075] Move job notifications to AppShell for always-on delivery
- Add JobNotification subscription to TendrilAppShell using UseEffect
- Remove notification hook from JobsApp (no longer needed)
- Delete NotificationHookDisposable method from JobsApp.Hooks.cs
- Notifications now fire regardless of which app is open
